refactor Dependency Injection to it namespace
