+++ src/Detection.Core/src/DependencyInjection/CoreCollectionExtensions.cs
@@ -2,14 +2,10 @@
 // The Apache v2. See License.txt in the project root for license information.
 
 using System;
-
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection.Extensions;
-
 using Wangkanai.Detection;
-using Wangkanai.Detection.Builders;
 
-// ReSharper disable once CheckNamespace
 namespace Microsoft.Extensions.DependencyInjection
 {
     /// <summary>
