+++ src/Detection.Core/src/DependencyInjection/DetectionBuilder.cs
@@ -5,7 +5,7 @@ using System;
 
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Wangkanai.Detection.Builders
+namespace Microsoft.Extensions.DependencyInjection
 {
     /// <summary>
     /// Helper functions for configuring detection services.
