+++ src/Detection.Core/src/DependencyInjection/IDetectionBuilder.cs
@@ -1,9 +1,7 @@
 // Copyright (c) 2014-2020 Sarin Na Wangkanai, All Rights Reserved.
 // The Apache v2. See License.txt in the project root for license information.
 
-using Microsoft.Extensions.DependencyInjection;
-
-namespace Wangkanai.Detection
+namespace Microsoft.Extensions.DependencyInjection
 {
     public interface IDetectionBuilder
     {
