+++ src/Detection.Core/src/Wangkanai.Detection.Core.csproj
@@ -22,8 +22,4 @@
     <FrameworkReference Include="Microsoft.AspNetCore.App" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="Builders\" />
-  </ItemGroup>
-
 </Project>
