+++ src/Responsive.Core/src/DependencyInjection/IResponsiveCoreBuilder.cs
@@ -1,9 +1,7 @@
-﻿// Copyright (c) 2014-2020 Sarin Na Wangkanai, All Rights Reserved.
+// Copyright (c) 2014-2020 Sarin Na Wangkanai, All Rights Reserved.
 // The Apache v2. See License.txt in the project root for license information.
 
-using Microsoft.Extensions.DependencyInjection;
-
-namespace Wangkanai.Responsive
+namespace Microsoft.Extensions.DependencyInjection
 {
     public interface IResponsiveCoreBuilder
     {
