+++ src/Responsive.Core/src/DependencyInjection/ResponsiveCoreBuilder.cs
@@ -1,11 +1,9 @@
-﻿// Copyright (c) 2014-2020 Sarin Na Wangkanai, All Rights Reserved.
+// Copyright (c) 2014-2020 Sarin Na Wangkanai, All Rights Reserved.
 // The Apache v2. See License.txt in the project root for license information.
 
 using System;
 
-using Microsoft.Extensions.DependencyInjection;
-
-namespace Wangkanai.Responsive.Configuration
+namespace Microsoft.Extensions.DependencyInjection
 {
     public class ResponsiveCoreBuilder : IResponsiveCoreBuilder
     {
