+++ src/Responsive.Core/test/ApplicationBuilderExtensionsTests.cs
@@ -4,6 +4,7 @@
 using System;
 
 using Microsoft.AspNetCore.Builder;
+using Microsoft.Extensions.DependencyInjection;
 using Xunit;
 
 namespace Wangkanai.Responsive.Core
